requirements.txt
discord.py
python_dotenv
discord-pretty-help
Discord-Anti-Spam
better_profanity
discord-pretty-help
pillow